refactor away from useCreateHook to renderHelper functions => less re-renders of the table content leading to higher performance
