+++ src/utils/table/index.js
@@ -1,4 +1,6 @@
+import renderTableHeader from './renderTableHeader';
+import renderTableData from './renderTableData';
 import downloadHelper from './downloadHelper';
 import determineToggle from './determineToggle';
 
-export { downloadHelper, determineToggle };
+export { renderTableHeader, renderTableData, downloadHelper, determineToggle };
